Measuring the position of a read/write head
The system includes a motor, a control system, and a sensor that measures the position of the read/write head. It analyses the system's transfer function, performance how external disturbances (like vibrations) affect the system's performance., as well as it simulates the system. Also, it designs a controller to improve the system's performance.